requirements.txt
adyen==3.0.0
amqp==2.6.1
aniso8601==7.0.0
asgiref==3.2.10
babel==2.8.0
beautifulsoup4==4.7.1
billiard==3.6.3.0
boto3==1.14.38
botocore==1.17.38
braintree==4.3.0
cachetools==4.1.1
cairocffi==1.1.0
cairosvg==2.4.2
celery==4.4.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-email-url==1.0.1
django==3.1
django-appconf==1.0.4
django-cache-url==3.1.2
django-countries==6.1.2
django-filter==2.3.0
django-js-asset==1.2.2
django-measurement==3.2.3
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-prices==2.2.0
django-prices-openexchangerates==1.1.0
django-prices-vatlayer==1.1.0
django-redis==4.12.1
django-render-block==0.7
django-storages==1.9.1
django-templated-email==2.3.0
django-versatileimagefield==2.0
docutils==0.15.2
draftjs-sanitizer==1.0.0
enmerkar==0.7.1
et-xmlfile==1.0.1
faker==4.1.1
freezegun==0.3.15
google-api-core==1.22.0
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-pubsub==1.7.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0; python_version >= "3.5"
google-i18n-address==2.3.5
google-measurement-protocol==1.0.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-django==2.13.0
graphene-federation==0.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpc-google-iam-v1==0.12.3
grpcio==1.31.0
gunicorn==20.0.4
h11==0.9.0
html-to-draftjs==1.0.1
html5lib==1.1
httptools==0.1.1; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
idna==2.10
jaeger-client==4.3.0
jdcal==1.4.1
jmespath==0.10.0
jsonfield==3.1.0
kombu==4.6.11
lxml==4.5.2
markdown==3.2.2
maxminddb==2.0.2
maxminddb-geolite2==2018.703
measurement==3.2.0
mpmath==1.1.0
oauthlib==3.1.0
openpyxl==3.0.4
opentracing==2.3.0
petl==1.6.2
phonenumberslite==8.12.7
pillow==7.2.0
prices==1.0.0
promise==2.3
protobuf==3.12.4
psycopg2-binary==2.8.5
purl==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyjwt==1.7.1
pyphen==0.9.5
python-dateutil==2.8.1
python-json-logger==0.1.11
python-magic==0.4.18
python-magic-bin==0.4.14; sys_platform == "win32"
pytimeparse==1.1.8
pytz==2020.1
razorpay==1.2.0
redis==3.5.3
requests==2.24.0
rsa==4.6; python_version >= "3.5"
rx==1.6.1
s3transfer==0.3.3
sentry-sdk==0.16.3
singledispatch==3.4.0.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.50.0
sympy==1.6.2
text-unidecode==1.3
threadloop==1.0.2
thrift==0.13.0
tinycss2==1.0.2
tornado==6.0.4
tqdm==4.48.2
typing==3.7.4.3
unidecode==1.1.1
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
vine==1.3.0
weasyprint==51
webencodings==0.5.1
websockets==8.1
django-taggit==1.3.0